requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
cycler==0.11.0
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.21.6
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
slim==0.1
torch==1.8.2+cu102
torchaudio==0.8.2
torchvision==0.9.2+cu102
typing_extensions==4.2.0
urllib3==1.26.9
